chore: rename from npm-manager to npm
