+++ extensions/npm/src/api/fetchPackageMetadata.ts
