+++ extensions/npm/src/components/NpmErrorDetails.tsx
