+++ extensions/npm/src/components/PackageDetails.tsx
