+++ extensions/npm/src/hooks/useInstallPackages.ts
