+++ extensions/npm/src/hooks/useUninstallPackages.ts
