+++ extensions/npm/src/hooks/useUpdatePackages.ts
