+++ extensions/npm/src/npm-install.tsx
