+++ extensions/npm/src/npm-uninstall.tsx
