+++ extensions/npm/src/utils/getPackageJson.ts
