+++ extensions/npm/src/utils/getPackageManager.ts
